Full screen

+++ ProjectFiles/detector.py
@@ -2,7 +2,8 @@ import cv2
 import time 
 import numpy as np
 import sqlite3
-from tkinter import *
+from Tkinter import * #change to Tkinter on PI
+import os
 
 def tableInsert(Id):
         name=0
@@ -23,54 +24,45 @@ def tableInsert(Id):
         return name
 
 
-faceDetect=cv2.CascadeClassifier('haarcascade_frontalface_default.xml');
-cam=cv2.VideoCapture(0);
-rec=cv2.face.LBPHFaceRecognizer_create();
-rec.read("recognizer/traningData.yml")
-font = cv2.FONT_HERSHEY_COMPLEX_SMALL
-t_end=time.time()+10
-name="Unknown"
 
+name="Unknown"
 sentinel=False
 Id=0
-#root = Tk()
 
-### timer
-##lbl1 = Label()
-##lbl1.pack(fill=BOTH, expand=1)
-##root.geometry("250x100+800+100")
-##lbl1.config(height=3, font=('times', 20, 'bold'))
+if os.path.exists("recognizer/traningData.yml"):
+        faceDetect=cv2.CascadeClassifier('haarcascade_frontalface_default.xml');
+        cam=cv2.VideoCapture(1); # change when on the Pi
+        rec=cv2.face.LBPHFaceRecognizer_create();
 
-while time.time()<=t_end:
-        cv2.namedWindow('image',cv2.WINDOW_NORMAL)
-        cv2.setWindowProperty("image",cv2.WND_PROP_FULLSCREEN,cv2.WINDOW_FULLSCREEN)
-       ## cv2.moveWindow("image", 0,0);
-        #cv2.resizeWindow('image', 480,320)
-        #cv2.setWindowProperty("image",WND_PROP_FULLSCREEN,WINDOW_FULLSCREEN);        
-        ret,img=cam.read();
-        gray=cv2.cvtColor(img,cv2.COLOR_BGR2GRAY);
-        faces=faceDetect.detectMultiScale(gray,1.3,5);
-        for(x,y,w,h) in faces:
-                cv2.rectangle(img,(x,y),(x+w,y+h),(0,255,0),2)
-                Id,conf=rec.predict(gray[y:y+h,x:x+w])
-                #print(conf);
-                if(conf<48):    # Play around with these values because camera is shitty
-                #print("confidence below 42")
-                        name=tableInsert(Id)
-                else:
-                        name="Unknown"
-                #cv2.putText(img,str(name),(x,y+h),font,3.0,(0,255,0));
+        rec.read("recognizer/traningData.yml")
+        font = cv2.FONT_HERSHEY_COMPLEX_SMALL
+        t_end=time.time()+10
+
+        while time.time()<=t_end:
+                cv2.namedWindow('image',cv2.WINDOW_NORMAL)
+                cv2.setWindowProperty("image",cv2.WND_PROP_FULLSCREEN,cv2.WINDOW_FULLSCREEN)
+                ret,img=cam.read();
+                gray=cv2.cvtColor(img,cv2.COLOR_BGR2GRAY);
+                faces=faceDetect.detectMultiScale(gray,1.3,5);
+                for(x,y,w,h) in faces:
+                        cv2.rectangle(img,(x,y),(x+w,y+h),(0,255,0),2)
+                        Id,conf=rec.predict(gray[y:y+h,x:x+w])
+                        #print(conf);
+                        if(conf<36):    # Play around with these values because camera is shitty
+                        #print("confidence below 42")
+                                name=tableInsert(Id)
+                        else:
+                                name="Unknown"
+                        #cv2.putText(img,str(name),(x,y+h),font,3.0,(0,255,0));
+                        
+                cv2.putText(img, "Done scanning in: " +str(int(t_end - time.time()))+" s",(280,35),font,1.25,(0,255,0));
+                cv2.imshow("image",img);
+                cv2.waitKey(1)
                 
-        cv2.putText(img, "Done scanning in: " +str(int(t_end - time.time()))+" s",(280,35),font,1.25,(0,255,0));
-        cv2.imshow("image",img);
-        cv2.waitKey(1)
-        #lbl1["text"] ="Done scanning in: "+ str(int(t_end - time.time()))
-        #root.update()
-             #   break;
-cam.release()
-cv2.destroyAllWindows()
-##root.destroy()
-##root.mainloop()
+        cam.release()
+        cv2.destroyAllWindows()
+        ##root.destroy()
+        ##root.mainloop()
 
 if name=="Unknown":
         #print "You are unauthorized to enter!"
